Extract protocol field filter codec
