+++ packages/protocol/src/protocol-query-codec.ts
@@ -12,6 +12,11 @@ import type {
 import { viewServerSchemaFieldMetadata } from "@view-server/config";
 import { Effect, Schema } from "effect";
 import {
+  decodeFilterValue,
+  encodeFilterValue,
+  type JsonFieldSchema,
+} from "./protocol-field-filter-codec";
+import {
   LooseWireGroupedQuerySchema,
   type LooseWireGroupedQuery,
   LooseWireRawQuerySchema,
@@ -20,8 +25,6 @@ import {
   type ViewServerWireGroupedQuery,
   type ViewServerWireRawQuery,
 } from "./protocol-query-schema";
-
-type JsonFieldSchema = Schema.Codec<unknown, unknown, never, never>;
 
 type TrustedRawQuery<Row> = {
   readonly select: ReadonlyArray<FieldKey<Row>>;
@@ -48,8 +51,6 @@ export type ViewServerValidatedLiveQuery<Row> =
   | ViewServerValidatedRawQuery<Row>
   | ViewServerValidatedGroupedQuery<Row>;
 
-const filterOperatorKeys = new Set(["eq", "neq", "in", "gt", "gte", "lt", "lte", "startsWith"]);
-const rangeFilterOperatorKeys = new Set(["gt", "gte", "lt", "lte"]);
 const dangerousRecordKeys = new Set(["__proto__", "prototype", "constructor"]);
 
 const strictParseOptions = {
@@ -62,11 +63,6 @@ const isRecord = (value: unknown): value is Record<string, unknown> =>
 const isNumericFieldSchema = (schema: JsonFieldSchema | undefined): boolean => {
   return schema !== undefined && viewServerSchemaFieldMetadata(schema).isNumeric;
 };
-
-const isFilterObject = (value: unknown): value is Record<string, unknown> =>
-  isRecord(value) &&
-  Object.keys(value).length > 0 &&
-  Object.keys(value).every((key) => filterOperatorKeys.has(key));
 
 const isGroupedQueryInput = (query: unknown): query is { readonly groupBy: unknown } =>
   isRecord(query) && Object.hasOwn(query, "groupBy");
@@ -154,180 +150,6 @@ export const viewServerDecodeHealthQuery = Effect.fn("ViewServerProtocol.healthQ
     return decoded;
   },
 );
-
-const encodeJsonFieldValue = Effect.fn("ViewServerProtocol.field.encode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  const encoded = yield* Schema.encodeUnknownEffect(Schema.toCodecJson(schema))(value).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Invalid filter for ${field}: ${error.message}`),
-    ),
-  );
-  return yield* Schema.decodeUnknownEffect(Schema.Json)(encoded).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Filter ${field} is not JSON-safe: ${error.message}`),
-    ),
-  );
-});
-
-const decodeJsonFieldValue = Effect.fn("ViewServerProtocol.field.decode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  return yield* Schema.decodeUnknownEffect(Schema.toCodecJson(schema))(value).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Invalid filter for ${field}: ${error.message}`),
-    ),
-  );
-});
-
-const encodeStringFilterValue = Effect.fn("ViewServerProtocol.filter.string.encode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  const decoded = yield* Schema.decodeUnknownEffect(Schema.toCodecJson(schema))(value).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Invalid filter for ${field}: ${error.message}`),
-    ),
-  );
-  if (typeof decoded !== "string") {
-    return yield* Effect.fail(invalidQuery(topic, `Filter ${field} does not support startsWith`));
-  }
-  return yield* Schema.decodeUnknownEffect(Schema.String)(value).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Invalid startsWith filter for ${field}: ${error.message}`),
-    ),
-  );
-});
-
-const decodeStringFilterValue = Effect.fn("ViewServerProtocol.filter.string.decode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  const decoded = yield* Schema.decodeUnknownEffect(Schema.toCodecJson(schema))(value).pipe(
-    Effect.mapError((error) =>
-      invalidQuery(topic, `Invalid filter for ${field}: ${error.message}`),
-    ),
-  );
-  if (typeof decoded !== "string") {
-    return yield* Effect.fail(invalidQuery(topic, `Filter ${field} does not support startsWith`));
-  }
-  return decoded;
-});
-
-const validateOperatorFilterValue = Effect.fn("ViewServerProtocol.filter.operator.validate")(
-  function* (
-    topic: string,
-    field: string,
-    schema: JsonFieldSchema,
-    value: Record<string, unknown>,
-  ) {
-    const metadata = viewServerSchemaFieldMetadata(schema);
-    if (metadata.isStructured) {
-      return;
-    }
-    const keys = Object.keys(value);
-    if (keys.includes("startsWith") && !metadata.isString) {
-      return yield* Effect.fail(invalidQuery(topic, `Filter ${field} does not support startsWith`));
-    }
-    if (keys.some((key) => rangeFilterOperatorKeys.has(key)) && !metadata.isNumeric) {
-      return yield* Effect.fail(
-        invalidQuery(topic, `Filter ${field} does not support range operators`),
-      );
-    }
-  },
-);
-
-const encodeOperatorFilterValue = Effect.fn("ViewServerProtocol.filter.operator.encode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: Record<string, unknown>,
-) {
-  yield* validateOperatorFilterValue(topic, field, schema, value);
-  const output: Record<string, Schema.Json> = {};
-  for (const [operator, operatorValue] of Object.entries(value)) {
-    if (operator === "in" && Array.isArray(operatorValue)) {
-      output[operator] = yield* Effect.forEach(operatorValue, (entry) =>
-        encodeJsonFieldValue(topic, field, schema, entry),
-      );
-    } else if (operator === "startsWith") {
-      output[operator] = yield* encodeStringFilterValue(topic, field, schema, operatorValue);
-    } else {
-      output[operator] = yield* encodeJsonFieldValue(topic, field, schema, operatorValue);
-    }
-  }
-  return output;
-});
-
-const decodeOperatorFilterValue = Effect.fn("ViewServerProtocol.filter.operator.decode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: Record<string, unknown>,
-) {
-  yield* validateOperatorFilterValue(topic, field, schema, value);
-  const output: Record<string, unknown> = {};
-  for (const [operator, operatorValue] of Object.entries(value)) {
-    if (operator === "in" && Array.isArray(operatorValue)) {
-      output[operator] = yield* Effect.forEach(operatorValue, (entry) =>
-        decodeJsonFieldValue(topic, field, schema, entry),
-      );
-    } else if (operator === "startsWith") {
-      output[operator] = yield* decodeStringFilterValue(topic, field, schema, operatorValue);
-    } else {
-      output[operator] = yield* decodeJsonFieldValue(topic, field, schema, operatorValue);
-    }
-  }
-  return output;
-});
-
-const encodeFilterValue = Effect.fn("ViewServerProtocol.filter.encode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  if (!isFilterObject(value)) {
-    return yield* encodeJsonFieldValue(topic, field, schema, value);
-  }
-  return yield* Effect.matchEffect(encodeOperatorFilterValue(topic, field, schema, value), {
-    onFailure: (operatorError) =>
-      Effect.matchEffect(encodeJsonFieldValue(topic, field, schema, value), {
-        onFailure: () => Effect.fail(operatorError),
-        onSuccess: Effect.succeed,
-      }),
-    onSuccess: Effect.succeed,
-  });
-});
-
-const decodeFilterValue = Effect.fn("ViewServerProtocol.filter.decode")(function* (
-  topic: string,
-  field: string,
-  schema: JsonFieldSchema,
-  value: unknown,
-) {
-  if (!isFilterObject(value)) {
-    return yield* decodeJsonFieldValue(topic, field, schema, value);
-  }
-  return yield* Effect.matchEffect(decodeOperatorFilterValue(topic, field, schema, value), {
-    onFailure: (operatorError) =>
-      Effect.matchEffect(decodeJsonFieldValue(topic, field, schema, value), {
-        onFailure: () => Effect.fail(operatorError),
-        onSuccess: Effect.succeed,
-      }),
-    onSuccess: Effect.succeed,
-  });
-});
 
 const encodeWhere = Effect.fn("ViewServerProtocol.query.where.encode")(function* <
   const Topics extends TopicDefinitions,
